fix(cli): isolate terminal title animation updates
Move terminal title frame cycling and stdout writes into a leaf component so the 10Hz spinner does not re-render the AppCoordinator/AppView tree.
